backend/admin_requirements.txt
fastapi-admin
sqlalchemy
aiosqlite
databases
jinja2
python-multipart
fastapi-users[sqlalchemy]
